+++ php/Exception/HydrationException.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace JsonQ\Exception;
+
+use RuntimeException;
+
+class HydrationException extends RuntimeException
+{
+}