+++ php/Exception/TypeMismatchException.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+namespace JsonQ\Exception;
+
+class TypeMismatchException extends HydrationException
+{
+}